Adding `immutable` to global load ops. (#16295)
This allows us to locally know whether a load is of an immutable global
or not. We only ever make globals immutable during FoldGlobals and
perform the fixup of all loads we can.

On programs with lots of globals this speeds things up quite a bit by
dropping propagateLiveness from e.g. 20s->12s (sdxl_turbo_unet).

+++ llvm-external-projects/iree-dialects/lib/Dialect/VectorExt/IR/VectorExtAttrs.cpp
@@ -35,8 +35,9 @@ std::optional<int64_t> PerDimLayoutAttr::getShape(const LayoutDimension &dim) {
 bool LayoutAttr::isValidLayout(ArrayRef<int64_t> shape) const {
   for (auto perDimLayout : llvm::enumerate(getLayouts())) {
     ArrayRef<int64_t> layoutShape = perDimLayout.value().getShapes();
-    int64_t computedShape = std::reduce(layoutShape.begin(), layoutShape.end(),
-                                        1, std::multiplies<int64_t>());
+    int64_t computedShape =
+        std::reduce(layoutShape.begin(), layoutShape.end(),
+                    static_cast<int64_t>(1), std::multiplies<int64_t>());
     int64_t expectedShape = shape[perDimLayout.index()];
     if (computedShape != expectedShape) {
       return false;
